Checks are made before deleting any folder.

+++ createDataFiles.m
@@ -1,6 +1,8 @@
 function createDataFiles(transVectors,centroids,V,nV,M)
-  rmdir('config','s');
-  mkdir('config');
+  if (exist('config','dir'))
+      rmdir('config','s');
+  end
+    mkdir('config');
   dlmwrite('config/transVec.dat',transVectors,','); 
   dlmwrite('config/centroids.dat',centroids,','); 
   dlmwrite('config/vecV.dat',V,',');
